fix(snapshot): enumeration immutability, third-table harness, dual-error evidence and test-split gate (review round 6)
1. Enumeration observations are truly immutable and carry provenance:
   counts become a plain Record (Object.freeze cannot freeze a Map's
   internal slots — the exact-head adversarial counts.set('feature',99)
   now throws), a required ENUM_HEAD section feeds the envelope's
   sourceRevision (no longer a hardcoded null), and a cache-poison
   regression asserts frozen reads across cache hits.

2. One contract-driven negative harness now consumes all THREE tables:
   the shared assertVariants helper runs missing/rc=128/garbage over
   worktree, repository and enumeration contracts. The parser enforces
   what the enumeration table declares — ENUM_BASE_AVAILABLE and
   ENUM_HEAD presence, and the counts section's presence and outer rc
   whenever the base is available (removing availability, removing
   counts, or rc=128 on the loop section all fail unknown now).

3. Enumeration retry preserves both raw attempts: the diagnostic
   records attempts:[first, second] with name/message/stack via the
   same describeSampleError evidence used by the repository advance.

4. The 853-line sampler test suite is split: the contract harness moved
   to tests/local-git-contract.test.ts and enumeration semantics to
   tests/local-git-enumeration.test.ts. The size gate's >800
   unconditional-split hard limit now covers tests/ too; the legacy
   4.4k-line routes suite is marked over800Legacy with a tracked
   dedicated split (#145) instead of silently passing. Stale sampler
   exception reasons updated to the post-split reality.

Review: https://github.com/ai-daming/clickvibe/pull/144#issuecomment-5461386776
Baseline: dc7be34. test 648/648; coverage 93.21/86.09/88.99; gates green.
