+++ tests/local-git-sampler.test.ts
@@ -586,268 +586,3 @@ test('a failed branch count is a required failure carrying the effective base re
   )
   assert.match(sample.requiredFailures[0].error, /fatal: bad revision/)
 })
-
-test('contract-driven negatives: every section reacts to missing/rc/garbage as classified', () => {
-  // Canonical healthy sample: worktree fully observable, base resolved via a
-  // named ref, upstream present, branch facts present.
-  const canonical: Array<[string, number, string]> = [
-    ['WT_GITDIR', 0, '/wt/.git'],
-    ['EB_NAMED', 0, 'origin/main'],
-    ['EB_COMPARE', 0, 'origin/main'],
-    ['EB_SOURCE', 0, 'named-ref'],
-    ['EB_AVAILABLE', 0, '1'],
-    ['WT_BASE_REF', 0, 'origin/main'],
-    ['WT_HEAD', 0, 'abc1234'],
-    ['WT_BRANCH', 0, 'feature'],
-    ['WT_STATUS', 0, ''],
-    ['WT_MAIN', 0, 'aaa0000'],
-    ['WT_MAIN_COUNT', 0, '0 1'],
-    ['WT_BASE', 0, 'bbb0000'],
-    ['WT_BASE_COUNT', 0, '0 2'],
-    ['WT_UPSTREAM', 0, 'abc1234'],
-    ['WT_UP_COUNT', 0, '0 0'],
-    ['WT_MERGE_HEAD', 1, ''],
-    ['BR_GITDIR', 0, '/repo/.git'],
-    ['BR_DEFAULT', 0, 'origin/main'],
-    ['BR_REF_ERROR', 0, ''],
-    ['BR_REF', 0, 'feature'],
-    ['BR_BASE_REF', 0, 'origin/main'],
-    ['BR_COMMIT_COUNT', 0, '2'],
-  ]
-  const build = (rows: Array<[string, number, string]>) =>
-    rows.map(([key, rc, value]) => section(key, rc, value)).join('\n')
-  const baseline = parseWorktreeSample(build(canonical))
-  assert.deepEqual(baseline.requiredFailures, [], 'canonical sample must be failure-free')
-
-  const verbFor = (entry: { section: string; producer: string }): string => {
-    if (entry.section.startsWith('EB_') || entry.section === 'WT_BASE_REF' || entry.section === 'BR_BASE_REF') {
-      return `effective-base ${entry.section}`
-    }
-    const verb = entry.producer.match(
-      /rev-parse --git-dir|rev-parse --short|rev-list --left-right|rev-list --count|show-ref|symbolic-ref|branch --show-current|status --porcelain|rev-parse/,
-    )
-    return verb ? verb[0].split(' ')[0] : entry.producer
-  }
-
-  for (const entry of WORKTREE_SECTION_CONTRACT) {
-    const index = canonical.findIndex(([key]) => key === entry.section)
-    if (index < 0) continue // conditional sections absent from this canonical sample
-
-    const variants: Array<{ name: string; make: () => string; expect: 'fail' | 'ok' | 'skip' }> = [
-      {
-        name: 'missing',
-        make: () => build(canonical.filter((_, i) => i !== index)),
-        expect: entry.negative.missing,
-      },
-      {
-        name: 'rc-failure',
-        make: () => {
-          const rows = canonical.map((row, i) =>
-            i === index ? ([row[0], 128, `fatal: ${row[0]} broke`] as [string, number, string]) : row,
-          )
-          return build(rows)
-        },
-        expect: entry.negative.rcNonZero,
-      },
-      {
-        name: 'garbage-value',
-        make: () => {
-          const rows = canonical.map((row, i) =>
-            i === index ? ([row[0], 0, 'garbage value'] as [string, number, string]) : row,
-          )
-          return build(rows)
-        },
-        expect: entry.negative.garbage,
-      },
-    ]
-    for (const variant of variants) {
-      if (variant.expect === 'skip') continue
-      const parsed = parseWorktreeSample(variant.make())
-      const verb = verbFor(entry)
-      const flagged = parsed.requiredFailures.some(
-        (failure) => failure.operation.includes(entry.section) || failure.operation.includes(verb),
-      )
-      assert.equal(
-        flagged,
-        variant.expect === 'fail',
-        `${entry.section} ${variant.name}: expected ${variant.expect}, operations=${JSON.stringify(parsed.requiredFailures.map((f) => f.operation))}`,
-      )
-    }
-  }
-})
-
-test('contract-driven negatives: repository sections and strict numeric shapes', () => {
-  const canonicalRepo: Array<[string, number, string]> = [
-    ['REPO_DEFAULT', 0, 'origin/main'],
-    ['REPO_BRANCH', 0, 'main'],
-    ['REPO_HEAD', 0, 'abc1234'],
-    ['REPO_MAIN_COUNT', 0, '0 1'],
-    ['REPO_HEAD_COUNT', 0, '0 1'],
-  ]
-  const buildRepo = (rows: Array<[string, number, string]>) =>
-    rows.map(([key, rc, value]) => section(key, rc, value)).join('\n')
-  assert.deepEqual(parseRepositorySample(buildRepo(canonicalRepo)).requiredFailures, [])
-
-  const verbFor = (entry: { section: string; producer: string }): string => {
-    const verb = entry.producer.match(/rev-parse|rev-list|symbolic-ref|branch --show-current/)
-    return verb ? verb[0].split(' ')[0] : entry.producer
-  }
-  for (const entry of REPOSITORY_SECTION_CONTRACT) {
-    const index = canonicalRepo.findIndex(([key]) => key === entry.section)
-    assert.ok(index >= 0, `${entry.section} must appear in the canonical repository sample`)
-    const variants: Array<{ name: string; make: () => string; expect: 'fail' | 'ok' | 'skip' }> = [
-      {
-        name: 'missing',
-        make: () => buildRepo(canonicalRepo.filter((_, i) => i !== index)),
-        expect: entry.negative.missing,
-      },
-      {
-        name: 'rc-failure',
-        make: () =>
-          buildRepo(
-            canonicalRepo.map((row, i) =>
-              i === index ? ([row[0], 128, 'fatal: broke'] as [string, number, string]) : row,
-            ),
-          ),
-        expect: entry.negative.rcNonZero,
-      },
-      {
-        name: 'garbage-value',
-        make: () =>
-          buildRepo(
-            canonicalRepo.map((row, i) =>
-              i === index ? ([row[0], 0, 'garbage value'] as [string, number, string]) : row,
-            ),
-          ),
-        expect: entry.negative.garbage,
-      },
-    ]
-    for (const variant of variants) {
-      if (variant.expect === 'skip') continue
-      const parsed = parseRepositorySample(variant.make())
-      const verb = verbFor(entry)
-      const flagged = parsed.requiredFailures.some(
-        (failure) => failure.operation.includes(entry.section) || failure.operation.includes(verb),
-      )
-      assert.equal(
-        flagged,
-        variant.expect === 'fail',
-        `${entry.section} ${variant.name}: expected ${variant.expect}, operations=${JSON.stringify(parsed.requiredFailures.map((f) => f.operation))}`,
-      )
-    }
-  }
-
-  const fractional = canonicalRepo.map((row) =>
-    row[0] === 'REPO_MAIN_COUNT' ? (['REPO_MAIN_COUNT', 0, '1.5 2'] as [string, number, string]) : row,
-  )
-  assert.ok(
-    parseRepositorySample(buildRepo(fractional)).requiredFailures.length > 0,
-    'a fractional compare value must fail',
-  )
-
-  const worktreeCanonical: Array<[string, number, string]> = [
-    ['WT_GITDIR', 0, '/wt/.git'],
-    ['EB_NAMED', 0, 'origin/main'],
-    ['EB_COMPARE', 0, 'origin/main'],
-    ['EB_SOURCE', 0, 'named-ref'],
-    ['EB_AVAILABLE', 0, '1'],
-    ['WT_BASE_REF', 0, 'origin/main'],
-    ['WT_HEAD', 0, 'abc1234'],
-    ['WT_BRANCH', 0, 'feature'],
-    ['WT_STATUS', 0, ''],
-    ['WT_MAIN', 1, ''],
-    ['WT_MAIN_COUNT', 1, ''],
-    ['WT_BASE', 0, 'bbb0000'],
-    ['WT_BASE_COUNT', 0, '1.5 2'],
-    ['WT_UPSTREAM', 127, ''],
-    ['WT_UP_COUNT', 127, ''],
-    ['WT_MERGE_HEAD', 1, ''],
-    ['BR_GITDIR', 0, '/repo/.git'],
-    ['BR_DEFAULT', 0, 'origin/main'],
-    ['BR_REF_ERROR', 0, ''],
-    ['BR_REF', 0, 'feature'],
-    ['BR_BASE_REF', 0, 'origin/main'],
-    ['BR_COMMIT_COUNT', 0, '1.5'],
-  ]
-  const parsed = parseWorktreeSample(worktreeCanonical.map(([key, rc, value]) => section(key, rc, value)).join('\n'))
-  assert.ok(
-    parsed.requiredFailures.some((failure) => failure.operation.includes('origin/main...HEAD')),
-    'a fractional base compare must fail',
-  )
-  assert.ok(
-    parsed.requiredFailures.some((failure) => failure.operation.includes('origin/main..feature')),
-    'a fractional commit count must fail, never hasCommits:true',
-  )
-  assert.equal(parsed.branchFacts.hasCommits, undefined)
-})
-
-test('repository enumeration: builder/contract/parser agree and counts stay strict', () => {
-  const command = buildRepositoryEnumerationCommand({ repoPath: '/repo/main' })
-  assert.match(command, /git -C '\/repo\/main' rev-parse --git-dir 2>&1/)
-  assert.match(
-    command,
-    /git -C '\/repo\/main' for-each-ref --format='%\(refname:short\)' refs\/heads refs\/remotes\/origin/,
-  )
-  assert.match(command, /rev-list --count "\$db\.\.\$br" 2>&1/)
-
-  const enc = (value: string) => Buffer.from(value, 'utf8').toString('base64')
-  const section = (key: string, rc: number, value: string) => `${key}\t${rc}\t${enc(value)}`
-  const healthy = [
-    section('ENUM_GITDIR', 0, '/repo/.git'),
-    section('ENUM_DEFAULT', 0, 'origin/main'),
-    section('ENUM_REFS', 0, 'main\norigin/main\nclickvibe-issue-122\norigin/clickvibe-issue-9'),
-    section('ENUM_BASE_AVAILABLE', 0, '1'),
-    section('ENUM_COUNTS', 0, 'main\t0\t0\nclickvibe-issue-122\t0\t2'),
-  ].join('\n')
-  const parsed = parseRepositoryEnumeration(healthy)
-  assert.deepEqual(parsed.requiredFailures, [])
-  assert.equal(parsed.defaultBranch, 'main')
-  assert.equal(parsed.counts.get('clickvibe-issue-122'), 2)
-  assert.ok(parsed.refs.includes('origin/clickvibe-issue-9'))
-
-  const noBase = [
-    section('ENUM_GITDIR', 0, '/repo/.git'),
-    section('ENUM_DEFAULT', 0, 'origin/main'),
-    section('ENUM_REFS', 0, 'main'),
-    section('ENUM_BASE_AVAILABLE', 0, '0'),
-    section('ENUM_COUNTS', 0, ''),
-  ].join('\n')
-  const noBaseParsed = parseRepositoryEnumeration(noBase)
-  assert.deepEqual(noBaseParsed.requiredFailures, [], 'deleted default base skips counts as expected absence')
-  assert.equal(noBaseParsed.counts.size, 0)
-
-  const failedCount = [
-    section('ENUM_GITDIR', 0, '/repo/.git'),
-    section('ENUM_DEFAULT', 0, 'origin/main'),
-    section('ENUM_REFS', 0, 'clickvibe-issue-122'),
-    section('ENUM_BASE_AVAILABLE', 0, '1'),
-    section('ENUM_COUNTS', 0, 'clickvibe-issue-122\t128\tfatal: bad object'),
-  ].join('\n')
-  const failed = parseRepositoryEnumeration(failedCount)
-  assert.deepEqual(
-    failed.requiredFailures.map((failure) => failure.operation),
-    ['git -C <repo> rev-list --count origin/main..clickvibe-issue-122'],
-  )
-  assert.match(failed.requiredFailures[0].error, /fatal: bad object/)
-
-  const fractional = [
-    section('ENUM_GITDIR', 0, '/repo/.git'),
-    section('ENUM_DEFAULT', 0, 'origin/main'),
-    section('ENUM_REFS', 0, 'clickvibe-issue-122'),
-    section('ENUM_BASE_AVAILABLE', 0, '1'),
-    section('ENUM_COUNTS', 0, 'clickvibe-issue-122\t0\t1.5'),
-  ].join('\n')
-  assert.ok(parseRepositoryEnumeration(fractional).requiredFailures.length > 0)
-
-  // static enumeration: builder sections ↔ contract table
-  const emitted = new Set([...command.matchAll(/'(ENUM_[A-Z_]+)\\t/g)].map((match) => match[1]))
-  const contracted = new Set(REPOSITORY_ENUMERATION_SECTION_CONTRACT.map((entry) => entry.section))
-  assert.deepEqual(
-    [...emitted].filter((name) => !contracted.has(name)),
-    [],
-  )
-  assert.deepEqual(
-    [...contracted].filter((name) => !emitted.has(name)),
-    [],
-  )
-})
